+++ src/core/renderer/index.d.ts
@@ -1,7 +1,6 @@
 // common
-export { TilesRendererBase } from './tiles/TilesRendererBase.js';
-export { Tile } from './tiles/Tile.js';
-export { TileBase } from './tiles/TileBase.js';
+export * from './tiles/TilesRendererBase.js';
+export { Tile, TileInternalData, TileTraversalData } from './tiles/Tile.js';
 export { Tileset } from './tiles/Tileset.js';
 export * from './loaders/B3DMLoaderBase.js';
 export * from './loaders/I3DMLoaderBase.js';