+++ CHANGELOG.md
@@ -4,14 +4,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.4.22] - Unreleased
+## [0.4.23] - Unreleased
+### Added
+- DebugTilesPlugin: Added support for "boundsColorMode".
+
+### Fixed
+- QuantizedMeshPlugin: Fixed case where child tiles could be added redundantly.
+- QuantizedMeshPlugin: Fixed case where tiles could throw error on disposal due to be incomplete. 
+- ImageOverlayPlugin: Fixed case where split tiles could be added and not be processed.
+
+## [0.4.22] - 2026-03-06
 ### Added
 - Types for gltf metadata extensions.
+- Event and other types to TilesRendererBase for use in thee.js and Babylon.js.
+- ImageOverlayPlugin: Add support for splitting "ADD" refinement tiles.
+- Babylon TilesRenderer: Added "checkCollisions" option.
 
 ### Fixed
 - LoadRegionPlugin: Add typings for new function, fix bug related to "distance" calculations.
 - GlobeContorls: Account for ellipsoid frame scale when calculating max zoom out distance.
 - Babylon TilesRenderer: Adjusted imports to support treeshaking.
+- ImageOverlayPlugin: Fix case where removing all overlays does not clear overlay textures correctly.
+- UnloadTilesPlugin: Fix case where the plugin would incorrectly retain tiles references, resulting in a memory leak.
+- OptimizedLoadStrategy Option: Add compatibility for QuantizedMeshPlugin, ImageOverlayPlugin.enableTileSplitting.
 
 ## [0.4.21] - 2026-02-04
 ### Fixed